Applied changes to log in
